requirements.txt
django
gunicorn
psycopg2-binary
celery
redis
python-dotenv
djangorestframework-simplejwt
djangorestframework
django-filter
requests
django-celery-beat
pandas